Split api-routes.ts into per-domain handler modules
handleApiRoute becomes a small dispatcher over a shared
ApiRouteContext built once per request. Each route group lives
in its own file under packages/server/src/http/api/:

  - context.ts          ApiRouteContext + buildApiRouteContext
  - admin.ts            handleAdminRoutes
  - application.ts      handleApplicationRoutes
  - topic-management.ts handleTopicManagementRoutes
  - artifacts.ts        handleArtifactRoutes
  - job-input.ts        handleJobInputRoutes
  - references.ts       handleReferenceRoutes
  - filesystem.ts       handleFilesystemRoutes

No behavior change, no payload-shape change, same branch
ordering as the previous monolith (47 routes preserved).
api-routes.ts shrinks from 1242 to 17 lines.
